+++ geotrek/core/tests/test_fields.py
@@ -3,11 +3,11 @@ from django.core.exceptions import ValidationError
 from django.contrib.gis.geos import LineString
 from django.conf import settings
 
-from geotrek.core.fields import SnappedLineStringField
+from geotrek.core.fields import SnappedLineStringField, TopologyField
 from geotrek.core.factories import PathFactory
 
 
-class FieldsTest(TestCase):
+class SnappedLineStringFieldTest(TestCase):
     def setUp(self):
         self.f = SnappedLineStringField()
         self.wktgeom = ('LINESTRING(-0.77054223313507 -5.32573853776343,'
@@ -18,22 +18,22 @@ class FieldsTest(TestCase):
 
     def test_dict_with_geom_is_mandatory(self):
         self.assertRaises(ValidationError, self.f.clean,
-                          ('LINESTRING(0 0, 1 0)',))
+                          'LINESTRING(0 0, 1 0)')
         self.assertRaises(ValidationError, self.f.clean,
-                          ('{"geo": "LINESTRING(0 0, 1 0)"}',))
+                          '{"geo": "LINESTRING(0 0, 1 0)"}')
 
     def test_snaplist_is_mandatory(self):
         self.assertRaises(ValidationError, self.f.clean,
-                          ('{"geom": "LINESTRING(0 0, 1 0)"}',))
+                          '{"geom": "LINESTRING(0 0, 1 0)"}')
 
     def test_snaplist_must_have_same_number_of_vertices(self):
         self.assertRaises(ValidationError, self.f.clean,
-                          ('{"geom": "LINESTRING(0 0, 1 0)", "snap": [null]}',))
+                          '{"geom": "LINESTRING(0 0, 1 0)", "snap": [null]}')
 
     def test_geom_cannot_be_invalid_wkt(self):
         self.assertRaises(ValidationError, self.f.clean,
-                          ('{"geom": "LINEPPRING(0 0, 1 0)", '
-                           '"snap": [null, null]}',))
+                          '{"geom": "LINEPPRING(0 0, 1 0)", '
+                          '"snap": [null, null]}')
 
     def test_geom_can_be_geojson(self):
         geojsonstr = self.geojson.replace('"', '\\"')
@@ -55,3 +55,11 @@ class FieldsTest(TestCase):
         self.assertTrue(self.f.clean(value).equals_exact(
             LineString((100000, 100000), (2, 2),
                        srid=settings.SRID), 0.1))
+
+
+class TopologyFieldTest(TestCase):
+    def setUp(self):
+        self.f = TopologyField()
+
+    def test_validation_fails_if_null_is_submitted(self):
+        self.assertRaises(ValidationError, self.f.clean, 'null')